Increase confidentiality guarantees for reporters
With this change, code of conduct violation reports will be only shared
with those who the reporter explicitly shares them with. Reports may
be sent to a subset of the committee if the reporter is uncomfortable
sharing the report with all members. Committee members' email addresses
are published to make them available for contact.